add changes in storage
