+++ straxen/storage/__init__.py
@@ -7,5 +7,8 @@ from .rucio import *
 from . import rundb
 from .rundb import *
 
+from .local_rucio import *
+from . import local_rucio
+
 from . import mongo_storage
 from .mongo_storage import *
